bulk_convert/requirements_full.txt
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0
# System deps required for OCR and pdf2image:
# - Tesseract OCR (install from https://github.com/tesseract-ocr/tesseract)
# - poppler (pdftoppm) for pdf2image (Windows: install poppler and add to PATH)
# Optional: LibreOffice for .doc -> .docx conversion (soffice on PATH)
